Add an extra topology type and do some code cleanup.
Add a topology for SMP / Hyperthreaded. Do these exist? It's not been tested
but should work if the core_id is 0 on all logical cpus.

Specify exit code 0 for sys.exit()'s in print_summary()

Use the number of logical cpu's found for the threads count rather than hard
coding it to 2.